pyproject.toml
[project]
name = "ocr-uv"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dashscope==1.23.2",
    "fastapi==0.104.0",
    "netifaces>=0.11.0",
    "pillow==10.0.1",
    "pydantic-settings>=2.9.1",
    "pymysql==1.1.1",
    "python-dotenv==1.0.0",
    "python-multipart==0.0.6",
    "uvicorn==0.23.2",
]
